requirements.txt
# Core
openai>=1.12.0
python-dotenv>=1.0.0

# NLP / ML
transformers>=4.38.0
torch>=2.1.0
scikit-learn>=1.3.0

# Data
pandas>=2.0.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.29.0

# Utilities
tqdm>=4.66.0
